+++ cli/cmd/dev/gerrit/static/entrypoint.sh
@@ -3,7 +3,7 @@
 set -ex
 
 # The /dev/./urandom is not a typo. https://stackoverflow.com/questions/58991966/what-java-security-egd-option-is-for
-JAVA_OPTIONS="-Djava.security.egd=file:/dev/./urandom"
+JAVA_OPTIONS="-Djava.security.egd=file:/dev/./urandom -Xms${JVM_XMS} -Xmx${JVM_XMX}"
 
 echo "Set local git config for gerrit admin"
 cat << EOF > ~/.gitconfig